add generic build script for all packages
