+++ package.json
@@ -5,9 +5,9 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "npm run build --workspaces",
-    "clean": "rimraf packages/**/lib",
-    "dev": "npm run build --workspaces --if-present",
+    "build": "tsx scripts/build.ts -p packages/graphql -p packages/typedoc",
+    "clean": "npx rimraf packages/**/lib",
+    "dev": "npm run build -- -w",
     "test": "npm test --workspaces"
   },
   "devDependencies": {
